refactor(ecstore): relocate NamespaceLocking + finalize God-Object split (backlog#822)
P7 (final) of the SetDisks God-Object split (tracking backlog#815, issue
backlog#822; depends on P6 #4290). Closes out the split:

1. NamespaceLocking归位: move impl NamespaceLocking for SetDisks out of
   set_disk/mod.rs and merge set_disk/lock.rs (lock formatting/mapping/online-
   disk helpers + tests) into a new set_disk/ops/locking.rs. set_disk/lock.rs is
   removed and 'mod lock;' dropped from mod.rs. Pure move — the NamespaceLocking
   contract stays implemented for SetDisks (associated-type bounds unchanged),
   and the lock helpers' pub(super) is widened to pub(in crate::set_disk) to
   preserve their exact prior set_disk-level reach from the deeper module.
   Token-stream diff of the moved code vs the originals is identical modulo
   visibility tokens and rustfmt signature re-wrapping.

2. Contract cleanup: remove the migration-era blanket #![allow(dead_code)] from
   storage_api_contracts/mod.rs and the residual List-contract placeholders it
   was masking (the never-wired-up pub(crate) EcstoreListOperations marker trait
   + blanket impl and its four type aliases). These were pub(crate), unreferenced
   anywhere in the crate (the contract-compat test derives its own copies from
   rustfs_storage_api directly), and inconsistent with the sibling re-export
   shims (bucket/heal/multipart). The contracts crate now compiles clean with
   dead_code enforced.

3. Documentation: add a set_disk module-layout doc comment to mod.rs describing
   the post-split structure (core, ctx, core::io_primitives, ops/* families,
   read.rs) and the backlog phase each piece came from.

Note: set_disk/mod.rs still carries its own #730 #![allow(dead_code/
unused_imports/unused_variables)] markers; those mask ~66 now-unused imports left
by the P5/P6 moves and are a larger, separate cleanup beyond this issue's stated
scope (which named storage_api_contracts).

Verification:
- cargo check / clippy -D warnings -p rustfs-ecstore --all-targets: clean
- cargo test -p rustfs-ecstore --lib: 1841 passed, 0 failed
- all five arch guard scripts (incl. doc-paths): pass
- storage_api_contracts compiles clean with the dead_code allow removed
