+++ crates/ecstore/src/set_disk/ops/mod.rs
@@ -20,5 +20,6 @@
 pub(crate) mod bucket;
 pub(crate) mod heal;
 pub(crate) mod list;
+pub(crate) mod locking;
 pub(crate) mod multipart;
 pub(crate) mod object;
